requirements.txt
elasticsearch[async]
discord.py==1.6.0
pyYAML
python-dateutil
aioipfs